1891 remove SaveMultipleDocumentsTask from U2Core and move it to U2Gui
