+++ src/corelibs/U2Core/src/tasks/SaveDocumentTask.h
@@ -83,6 +83,9 @@ public:
     /** Sets new 'openDocumentWithProjectHints'. See 'openDocumentWithProjectHints' for details. */
     void setOpenDocumentWithProjectHints(const QVariantMap& hints);
 
+    // Returns true if there is no write permission for the given url.
+    static bool isNoWritePermission(GUrl& url);
+
 private:
     StateLock* lock = nullptr;
     QPointer<Document> doc;
@@ -98,17 +101,6 @@ private:
 enum SavedNewDocFlag {
     SavedNewDoc_Open = true,
     SavedNewDoc_DoNotOpen = false
-};
-
-class U2CORE_EXPORT SaveMultipleDocuments : public Task {
-    Q_OBJECT
-public:
-    SaveMultipleDocuments(const QList<Document*>& docs, bool askBeforeSave, SavedNewDocFlag openFlag = SavedNewDoc_DoNotOpen);
-
-    static QList<Document*> findModifiedDocuments(const QList<Document*>& docs);
-
-private:
-    GUrl chooseAnotherUrl(Document* doc);
 };
 
 class U2CORE_EXPORT SaveCopyAndAddToProjectTask : public Task {
