+++ src/corelibs/U2View/src/ov_msa/CreateSubalignmentDialogController.cpp
@@ -27,7 +27,6 @@
 
 #include <U2Algorithm/CreateSubalignmentTask.h>
 
-#include <U2Core/AddDocumentTask.h>
 #include <U2Core/AppContext.h>
 #include <U2Core/BaseDocumentFormats.h>
 #include <U2Core/DocumentModel.h>
@@ -38,6 +37,7 @@
 
 #include <U2Formats/GenbankLocationParser.h>
 
+#include <U2Gui/AddDocumentTask.h>
 #include <U2Gui/HelpButton.h>
 #include <U2Gui/OpenViewTask.h>
 #include <U2Gui/SaveDocumentController.h>
