--- src/include/U2Core/AddDocumentTask.h
@@ -1,1 +1,0 @@
-#include "../../corelibs/U2Core/src/tasks/AddDocumentTask.h"
