--- src/include/U2Core/RemoveDocumentTask.h
@@ -1,1 +1,0 @@
-#include "../../corelibs/U2Core/src/tasks/RemoveDocumentTask.h"
