+++ src/include/U2Gui/AddDocumentTask.h
@@ -1,0 +1,1 @@
+#include "../../corelibs/U2Gui/src/AddDocumentTask.h"
