+++ src/include/U2Gui/RemoveDocumentTask.h
@@ -1,0 +1,1 @@
+#include "../../corelibs/U2Gui/src/RemoveDocumentTask.h"
