+++ src/include/U2Gui/SaveMultipleDocumentTask.h
@@ -1,0 +1,1 @@
+#include "../../corelibs/U2Gui/src/SaveMultipleDocumentTask.h"
