+++ src/plugins/circular_view/src/ShiftSequenceStartTask.cpp
@@ -21,7 +21,6 @@
 
 #include "ShiftSequenceStartTask.h"
 
-#include <U2Core/AddDocumentTask.h>
 #include <U2Core/AnnotationTableObject.h>
 #include <U2Core/AppContext.h>
 #include <U2Core/Counter.h>
@@ -35,6 +34,8 @@
 #include <U2Core/U1AnnotationUtils.h>
 #include <U2Core/U2SafePoints.h>
 #include <U2Core/U2SequenceUtils.h>
+
+#include <U2Gui/AddDocumentTask.h>
 
 namespace U2 {
